backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
blis==0.7.11
cachetools==5.5.1
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.16.0
confection==0.1.5
cymem==2.0.11
diskcache==5.6.3
fastapi==0.104.1
filelock==3.17.0
fsspec==2025.2.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.63.0rc1
h11==0.14.0
huggingface-hub==0.17.3
idna==3.10
Jinja2==3.1.5
langcodes==3.5.0
language_data==1.3.0
llama_cpp_python==0.3.7
marisa-trie==1.2.1
MarkupSafe==3.0.2
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
numpy==2.2.3
packaging==24.2
preshed==3.0.9
proto-plus==1.26.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.2
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.22
srsly==2.5.1
starlette==0.27.0
sympy==1.13.1
thinc==8.2.5
tokenizers==0.14.1
torch==2.6.0
tqdm==4.67.1
transformers==4.34.0
typer==0.9.4
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.24.0
wasabi==1.1.3
weasel==0.3.4
websockets==11.0.3
